Prep for workspace: move hook types to config, decouple error from reqwest
- HookEvent, HookAction, HookDefinition moved from hooks/mod.rs to
  config/schema.rs (hooks module re-exports them)
- LlmError::Http changed from reqwest::Error to String
- client.rs uses .map_err() for reqwest errors
- This removes reqwest dependency from error.rs and config

No functional changes. All 182 tests pass.
